add submit question

+++ src/main/java/learn/tassel/community/Controller/PublishController.java
@@ -32,7 +32,7 @@ public class PublishController {
     }
 
     @GetMapping("/publish/{id}")
-    public String editPublish(@PathVariable(name = "id")Integer id,
+    public String editPublish(@PathVariable(name = "id")Long id,
                               Model model){
         QuestionDTO question = questionService.findQuestionById(id);
 
@@ -49,7 +49,7 @@ public class PublishController {
     public String doPublish(@RequestParam("title")String title,
                             @RequestParam("description")String description,
                             @RequestParam("tag")String tag,
-                            @RequestParam("id") Integer id,
+                            @RequestParam("id") Long id,
                             HttpServletRequest request,
                             Model model){
