+++ src/main/java/learn/tassel/community/Model/Question.java
@@ -4,7 +4,7 @@ import lombok.Data;
 
 @Data
 public class Question {
-    private Integer id;//发布信息id
+    private Long id;//发布信息id
     private String title;
     private String description;
     private Long gmtCreate;
